+++ admin/views/layout-footer.php
@@ -42,6 +42,7 @@ if (!isset($kernel)) {
 <script src="<?= esc_url(admin_asset_url('js/widget-editor.js')) ?>" defer></script>
 <script src="<?= esc_url(admin_asset_url('js/theme-file-editor.js')) ?>" defer></script>
 <script src="<?= esc_url(admin_asset_url('js/folder-drag-drop.js')) ?>" defer></script>
+<script src="<?= esc_url(admin_asset_url('js/folder-tree-toggle.js')) ?>" defer></script>
 <script src="<?= esc_url(admin_asset_url('js/sortable.js')) ?>" defer></script>
 <script src="<?= esc_url(admin_asset_url('js/quick-edit.js')) ?>" defer></script>
 <script src="<?= esc_url(admin_asset_url('js/update-upload.js')) ?>" defer></script>